File Tree Setup
Added various key files based on what we used on the last project. I
don't mind having all these files in the main branch, there are,
however, two image folders are set up. The 'images' folder is used for
images, it's a photography website. The 'UI_images' folder is for those
images that are required for the user interface (i.e. logos and stuff)
I can't imagine that this folder is going to be rather large at all.